reorganize auth components and hooks
