+++ site/src/components/Dashboard/DashboardLayout.tsx
@@ -4,10 +4,10 @@ import Button from "@mui/material/Button";
 import InfoOutlined from "@mui/icons-material/InfoOutlined";
 import { type FC, type HTMLAttributes, Suspense } from "react";
 import { Outlet } from "react-router-dom";
+import { usePermissions } from "contexts/auth/usePermissions";
 import { LicenseBanner } from "components/Dashboard/LicenseBanner/LicenseBanner";
 import { Loader } from "components/Loader/Loader";
 import { ServiceBanner } from "components/Dashboard/ServiceBanner/ServiceBanner";
-import { usePermissions } from "hooks/usePermissions";
 import { dashboardContentBottomPadding } from "theme/constants";
 import { docs } from "utils/docs";
 import { Navbar } from "./Navbar/Navbar";
