+++ site/src/components/Dashboard/DeploymentBanner/DeploymentBanner.tsx
@@ -1,9 +1,9 @@
 import { type FC } from "react";
 import { useQuery } from "react-query";
-import { deploymentStats } from "api/queries/deployment";
-import { usePermissions } from "hooks/usePermissions";
-import { DeploymentBannerView } from "./DeploymentBannerView";
 import { health } from "api/queries/debug";
+import { deploymentStats } from "api/queries/deployment";
+import { usePermissions } from "contexts/auth/usePermissions";
+import { DeploymentBannerView } from "./DeploymentBannerView";
 
 export const DeploymentBanner: FC = () => {
   const permissions = usePermissions();
