+++ site/src/contexts/auth/RequirePermission.tsx
@@ -1,8 +1,8 @@
-import { FC } from "react";
+import { type FC, type ReactNode } from "react";
 import { Navigate } from "react-router-dom";
 
 export interface RequirePermissionProps {
-  children: JSX.Element;
+  children?: ReactNode;
   isFeatureVisible: boolean;
 }
 
@@ -16,6 +16,6 @@ export const RequirePermission: FC<RequirePermissionProps> = ({
   if (!isFeatureVisible) {
     return <Navigate to="/workspaces" />;
   } else {
-    return children;
+    return <>{children}</>;
   }
 };
