+++ site/src/hooks/index.ts
@@ -2,9 +2,5 @@ export * from "./useClickable";
 export * from "./useClickableTableRow";
 export * from "./useClipboard";
 export * from "./useFeatureVisibility";
-export * from "./useLocalStorage";
-export * from "./useMe";
-export * from "./useOrganizationId";
 export * from "./usePagination";
-export * from "./usePermissions";
 export * from "./useTab";
