--- site/src/hooks/useOrganizationId.ts
@@ -1,6 +1,0 @@
-import { useMe } from "./useMe";
-
-export const useOrganizationId = (): string => {
-  const user = useMe();
-  return user.organization_ids[0];
-};
