+++ site/src/pages/CreateUserPage/CreateUserPage.tsx
@@ -1,13 +1,13 @@
-import { useOrganizationId } from "hooks/useOrganizationId";
-import { FC } from "react";
+import { type FC } from "react";
 import { Helmet } from "react-helmet-async";
-import { useNavigate } from "react-router-dom";
-import { CreateUserForm } from "./CreateUserForm";
-import { Margins } from "components/Margins/Margins";
-import { pageTitle } from "utils/page";
 import { useMutation, useQuery, useQueryClient } from "react-query";
+import { useNavigate } from "react-router-dom";
+import { pageTitle } from "utils/page";
 import { authMethods, createUser } from "api/queries/users";
+import { useOrganizationId } from "contexts/auth/useOrganizationId";
+import { Margins } from "components/Margins/Margins";
 import { displaySuccess } from "components/GlobalSnackbar/utils";
+import { CreateUserForm } from "./CreateUserForm";
 
 export const Language = {
   unknownError: "Oops, an unknown error occurred.",
