+++ site/src/pages/ExternalAuthPage/ExternalAuthPage.tsx
@@ -1,18 +1,18 @@
+import Button from "@mui/material/Button";
 import { useQuery, useQueryClient } from "react-query";
-import { usePermissions } from "hooks";
+import { isAxiosError } from "axios";
 import { type FC } from "react";
 import { useParams, useSearchParams } from "react-router-dom";
-import ExternalAuthPageView from "./ExternalAuthPageView";
 import { ApiErrorResponse } from "api/errors";
-import { isAxiosError } from "axios";
-import Button from "@mui/material/Button";
-import { SignInLayout } from "components/SignInLayout/SignInLayout";
-import { Welcome } from "components/Welcome/Welcome";
 import {
   externalAuthDevice,
   externalAuthProvider,
   exchangeExternalAuthDevice,
 } from "api/queries/externalAuth";
+import { usePermissions } from "contexts/auth/usePermissions";
+import { SignInLayout } from "components/SignInLayout/SignInLayout";
+import { Welcome } from "components/Welcome/Welcome";
+import ExternalAuthPageView from "./ExternalAuthPageView";
 
 const ExternalAuthPage: FC = () => {
   const { provider } = useParams() as { provider: string };
