+++ site/src/pages/GroupsPage/CreateGroupPage.tsx
@@ -1,11 +1,11 @@
-import { useOrganizationId } from "hooks/useOrganizationId";
-import { FC } from "react";
+import { type FC } from "react";
 import { Helmet } from "react-helmet-async";
+import { useMutation, useQueryClient } from "react-query";
 import { useNavigate } from "react-router-dom";
+import { createGroup } from "api/queries/groups";
+import { useOrganizationId } from "contexts/auth/useOrganizationId";
 import { pageTitle } from "utils/page";
 import CreateGroupPageView from "./CreateGroupPageView";
-import { useMutation, useQueryClient } from "react-query";
-import { createGroup } from "api/queries/groups";
 
 export const CreateGroupPage: FC = () => {
   const queryClient = useQueryClient();
