+++ site/src/pages/GroupsPage/GroupsPage.tsx
@@ -1,14 +1,14 @@
-import { useFeatureVisibility } from "hooks/useFeatureVisibility";
-import { useOrganizationId } from "hooks/useOrganizationId";
-import { usePermissions } from "hooks/usePermissions";
-import { FC, useEffect } from "react";
+import { type FC, useEffect } from "react";
 import { Helmet } from "react-helmet-async";
-import { pageTitle } from "utils/page";
-import GroupsPageView from "./GroupsPageView";
 import { useQuery } from "react-query";
-import { groups } from "api/queries/groups";
-import { displayError } from "components/GlobalSnackbar/utils";
 import { getErrorMessage } from "api/errors";
+import { groups } from "api/queries/groups";
+import { useOrganizationId } from "contexts/auth/useOrganizationId";
+import { usePermissions } from "contexts/auth/usePermissions";
+import { useFeatureVisibility } from "hooks/useFeatureVisibility";
+import { pageTitle } from "utils/page";
+import { displayError } from "components/GlobalSnackbar/utils";
+import GroupsPageView from "./GroupsPageView";
 
 export const GroupsPage: FC = () => {
   const organizationId = useOrganizationId();
