+++ site/src/pages/StarterTemplatePage/StarterTemplatePage.tsx
@@ -1,11 +1,11 @@
-import { useOrganizationId } from "hooks/useOrganizationId";
-import { FC } from "react";
+import { type FC } from "react";
 import { Helmet } from "react-helmet-async";
-import { useParams } from "react-router-dom";
-import { pageTitle } from "utils/page";
-import { StarterTemplatePageView } from "./StarterTemplatePageView";
 import { useQuery } from "react-query";
+import { useParams } from "react-router-dom";
 import { templateExamples } from "api/queries/templates";
+import { pageTitle } from "utils/page";
+import { useOrganizationId } from "contexts/auth/useOrganizationId";
+import { StarterTemplatePageView } from "./StarterTemplatePageView";
 
 const StarterTemplatePage: FC = () => {
   const { exampleId } = useParams() as { exampleId: string };
