+++ site/src/pages/StarterTemplatesPage/StarterTemplatesPage.tsx
@@ -1,11 +1,11 @@
-import { useOrganizationId } from "hooks/useOrganizationId";
-import { FC } from "react";
+import { type FC } from "react";
 import { Helmet } from "react-helmet-async";
-import { pageTitle } from "utils/page";
-import { StarterTemplatesPageView } from "./StarterTemplatesPageView";
 import { useQuery } from "react-query";
 import { templateExamples } from "api/queries/templates";
+import { useOrganizationId } from "contexts/auth/useOrganizationId";
+import { pageTitle } from "utils/page";
 import { getTemplatesByTag } from "utils/starterTemplates";
+import { StarterTemplatesPageView } from "./StarterTemplatesPageView";
 
 const StarterTemplatesPage: FC = () => {
   const organizationId = useOrganizationId();
