+++ site/src/pages/TemplatePage/TemplateFilesPage/TemplateFilesPage.tsx
@@ -1,15 +1,15 @@
 import { type FC } from "react";
 import { Helmet } from "react-helmet-async";
+import { useQuery } from "react-query";
+import { previousTemplateVersion, templateFiles } from "api/queries/templates";
 import { Loader } from "components/Loader/Loader";
 import {
   TemplateFiles,
   useFileTab,
 } from "components/TemplateFiles/TemplateFiles";
 import { useTemplateLayoutContext } from "pages/TemplatePage/TemplateLayout";
+import { useOrganizationId } from "contexts/auth/useOrganizationId";
 import { getTemplatePageTitle } from "../utils";
-import { useQuery } from "react-query";
-import { previousTemplateVersion, templateFiles } from "api/queries/templates";
-import { useOrganizationId } from "hooks";
 
 const TemplateFilesPage: FC = () => {
   const orgId = useOrganizationId();
