+++ site/src/pages/TemplateSettingsPage/TemplateSettingsLayout.tsx
@@ -1,16 +1,16 @@
 import { createContext, type FC, Suspense, useContext } from "react";
 import { Helmet } from "react-helmet-async";
 import { useQuery } from "react-query";
-import { pageTitle } from "utils/page";
-import { Stack } from "components/Stack/Stack";
-import { Loader } from "components/Loader/Loader";
 import { Outlet, useParams } from "react-router-dom";
-import { Margins } from "components/Margins/Margins";
-import { useOrganizationId } from "hooks/useOrganizationId";
+import { pageTitle } from "utils/page";
+import { useOrganizationId } from "contexts/auth/useOrganizationId";
+import { checkAuthorization } from "api/queries/authCheck";
 import { templateByName } from "api/queries/templates";
 import type { AuthorizationResponse, Template } from "api/typesGenerated";
 import { ErrorAlert } from "components/Alert/ErrorAlert";
-import { checkAuthorization } from "api/queries/authCheck";
+import { Loader } from "components/Loader/Loader";
+import { Margins } from "components/Margins/Margins";
+import { Stack } from "components/Stack/Stack";
 import { Sidebar } from "./Sidebar";
 
 const TemplateSettings = createContext<
