+++ site/src/pages/TemplateVersionEditorPage/TemplateVersionEditorPage.tsx
@@ -1,10 +1,20 @@
-import { TemplateVersionEditor } from "./TemplateVersionEditor";
-import { useOrganizationId } from "hooks/useOrganizationId";
-import { FC, useEffect, useState } from "react";
+import { type FC, useEffect, useState } from "react";
 import { Helmet } from "react-helmet-async";
-import { useNavigate, useParams } from "react-router-dom";
-import { pageTitle } from "utils/page";
 import { useMutation, useQuery, useQueryClient } from "react-query";
+import { useNavigate, useParams } from "react-router-dom";
+import { TemplateVersionEditor } from "./TemplateVersionEditor";
+import { useOrganizationId } from "contexts/auth/useOrganizationId";
+import { pageTitle } from "utils/page";
+import {
+  patchTemplateVersion,
+  updateActiveTemplateVersion,
+  watchBuildLogsByTemplateVersionId,
+} from "api/api";
+import type {
+  PatchTemplateVersionRequest,
+  ProvisionerJobLog,
+  TemplateVersion,
+} from "api/typesGenerated";
 import {
   createTemplateVersion,
   resources,
@@ -19,16 +29,6 @@ import {
   createTemplateVersionFileTree,
   isAllowedFile,
 } from "utils/templateVersion";
-import {
-  patchTemplateVersion,
-  updateActiveTemplateVersion,
-  watchBuildLogsByTemplateVersionId,
-} from "api/api";
-import {
-  PatchTemplateVersionRequest,
-  ProvisionerJobLog,
-  TemplateVersion,
-} from "api/typesGenerated";
 import { displayError } from "components/GlobalSnackbar/utils";
 import { FullScreenLoader } from "components/Loader/FullScreenLoader";
