+++ site/src/pages/UsersPage/UsersLayout.tsx
@@ -2,12 +2,12 @@ import Button from "@mui/material/Button";
 import Link from "@mui/material/Link";
 import GroupAdd from "@mui/icons-material/GroupAddOutlined";
 import PersonAdd from "@mui/icons-material/PersonAddOutlined";
+import { type FC, Suspense } from "react";
+import { Link as RouterLink, Outlet, useNavigate } from "react-router-dom";
 import { USERS_LINK } from "components/Dashboard/Navbar/NavbarView";
 import { PageHeader, PageHeaderTitle } from "components/PageHeader/PageHeader";
 import { useFeatureVisibility } from "hooks/useFeatureVisibility";
-import { usePermissions } from "hooks/usePermissions";
-import { FC, Suspense } from "react";
-import { Link as RouterLink, Outlet, useNavigate } from "react-router-dom";
+import { usePermissions } from "contexts/auth/usePermissions";
 import { Margins } from "components/Margins/Margins";
 import { TabLink, Tabs } from "components/Tabs/Tabs";
 import { Loader } from "components/Loader/Loader";
