+++ site/src/testHelpers/renderHelpers.tsx
@@ -9,7 +9,7 @@ import { QueryClient } from "react-query";
 import { AppProviders } from "App";
 import { ThemeProvider } from "contexts/ThemeProvider";
 import { DashboardLayout } from "components/Dashboard/DashboardLayout";
-import { RequireAuth } from "components/RequireAuth/RequireAuth";
+import { RequireAuth } from "contexts/auth/RequireAuth";
 import { TemplateSettingsLayout } from "pages/TemplateSettingsPage/TemplateSettingsLayout";
 import { WorkspaceSettingsLayout } from "pages/WorkspaceSettingsPage/WorkspaceSettingsLayout";
 import {
